Add default.cfg for P2 mouse inputs
MAME's SDL input module (input_sdl.cpp) is hard-coded to support
only one keyboard and one mouse, claiming it's an SDL limitation.

That was true in 2016 when the code comments were last touched,
but unclear if it's still the case. Still investigating SDL.

In any case, for now player 1 and 2 will have to share a single
mouse for trackball and dial inputs. This is undesirable, however,
because each player can interfere with the opposing player's game.