attract.cfg: Exclude Centipede/Millipede from 8-way list
Sometime in the last year or so, MAME's XML output added an 8-way
joystick to the controls list for Centipede and Millipede causing
these games to be listed in both the "8-way" and "trackball" game
lists for Attract Mode.

We obviously don't want these trackball games in the "8-way" game
list so add a rule to exclude games with a trackball control.